requirements.txt
pandas>=2.1,<2.3
numpy>=1.24,<2.0
scikit-learn
requests
SQLAlchemy
psycopg2-binary
python-dotenv
streamlit
